Refactor: move Ray scheduler into RayEngine subclass
- Move _launch_workers and _launch_scheduler_processes from module-level
  functions into Engine instance methods, making them overridable
- Introduce RayEngine(Engine) subclass in sglang.srt.ray.engine that
  overrides _launch_scheduler_processes to launch Ray actors
- Extract _setup_and_run_http_server from launch_server() in http_server.py
  so both mp and Ray launchers can share the HTTP setup logic
- Add sglang.srt.ray.http_server with launch_ray_server() for --use-ray
- Route --use-ray in launch_server.py to the new Ray HTTP server
- Delete sglang.srt.ray.launch (replaced by RayEngine)
- Rename _launch_workers to _launch_subprocesses for backward compat shim
