+++ python/sglang/srt/ray/__init__.py
@@ -1,0 +1,3 @@
+from sglang.srt.ray.engine import RayEngine
+
+__all__ = ["RayEngine"]
