+++ python/sglang/srt/ray/http_server.py
@@ -1,0 +1,61 @@
+# Copyright 2023-2024 SGLang Team
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+# ==============================================================================
+"""Ray-aware HTTP server launcher."""
+
+from typing import Callable, Optional
+
+from sglang.srt.server_args import ServerArgs
+
+
+def _launch_ray_subprocesses(server_args: ServerArgs):
+    """Launch subprocesses using RayEngine (Ray actor scheduler backend)."""
+    from sglang.srt.ray.engine import RayEngine
+
+    launcher = object.__new__(RayEngine)
+    return launcher._launch_workers(server_args=server_args)
+
+
+def launch_ray_server(
+    server_args: ServerArgs,
+    execute_warmup_func: Optional[Callable] = None,
+    launch_callback: Optional[Callable[[], None]] = None,
+):
+    """Launch HTTP server with Ray-based scheduler actors.
+
+    Mirrors http_server.launch_server() but uses RayEngine for scheduler launching.
+    """
+    from sglang.srt.entrypoints.http_server import (
+        _execute_server_warmup,
+        _setup_and_run_http_server,
+    )
+
+    if execute_warmup_func is None:
+        execute_warmup_func = _execute_server_warmup
+
+    (
+        tokenizer_manager,
+        template_manager,
+        port_args,
+        scheduler_result,
+    ) = _launch_ray_subprocesses(server_args)
+
+    _setup_and_run_http_server(
+        server_args,
+        tokenizer_manager,
+        template_manager,
+        port_args,
+        scheduler_result.scheduler_infos,
+        execute_warmup_func=execute_warmup_func,
+        launch_callback=launch_callback,
+    )
